datagen/requirements.txt
numpy==2.2.4
opencv-python==4.11.0.86
pycocotools==2.0.8
scikit-learn==1.6.1
openai==1.66.5
torch
kagglehub
sentencepiece
absl-py
transformers
accelerate>=0.26.0
tiktoken
pandas
flask
matplotlib>=3.7.1
opencv-python
google-cloud-storage
vertexai
dotenv
google-genai
seaborn>=0.12.2
numpy>=1.24.3
setproctitle
scikit-image
kaggle